process SIGTERM in sleep-infinity script
sh queues signals while other commands are executing, and processes the
signals between commands. So if we were sleeping for 10000d we had to
wait all this time to be able to process SIGTERM. So change the sleep
interval to something more bearable (it should be lower than 10 seconds
(time between docker sending SIGTERM and SIGKILL when doing `docker
stop`) and enough let other processes in the container to stop).